docs(products): Phase 1 sync verification + runner script + schema fix
Records the observed row counts, label coverage, FK integrity, and
idempotency from the dev-Supabase sync after the schema split.

Schema fix: status_code_id and tax_type_override_id were SMALLINT in the
original product_inventory migration, but Prism's StatusCodeID can hold
32-bit values (observed 1246576928 as a sentinel). Widened both columns to
INTEGER via migration 20260419182000_fix_product_inventory_id_types and
applied the ALTER to dev Supabase before the successful sync run.

Sync results (first run): scanned=97638, updated=159051, removed=0, 60.7s
Row counts: products=61395, PIER=61024, PCOP=18511, PFS=18103, PBO=0
Second run idempotent: same scanned/updated, removed=0, counts stable.

scripts/run-phase-1-sync.ts is a minimal CLI runner that invokes
runPrismPull directly — used during verification and available for
future re-runs when the UI "Sync Database" button isn't handy.

+++ prisma/migrations/20260419180000_product_inventory_table/migration.sql
@@ -20,9 +20,9 @@ CREATE TABLE IF NOT EXISTS "product_inventory" (
   -- Label columns (denormalized from Prism ref tables each sync)
   "tag_type_id"             INTEGER,
   "tag_type_label"          TEXT,
-  "status_code_id"          SMALLINT,
+  "status_code_id"          INTEGER,
   "status_code_label"       TEXT,
-  "tax_type_override_id"    SMALLINT,
+  "tax_type_override_id"    INTEGER,
   "disc_code_id"            INTEGER,
 
   -- Reorder / stocking controls
